requirements.txt
SQLAlchemy==2.0.36
psycopg2-binary==2.9.6
gunicorn==23.0.0
python-telegram-bot==20.3
cloudinary==1.30.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Session
python-dotenv 
